requirements.txt
streamlit==1.31.0
groq==0.4.1
langgraph==0.0.26
langchain==0.1.6
langchain-openai==0.0.5
plotly==5.18.0
python-dotenv==1.0.0

